+++ src/check_python_versions/sources/tox.py
@@ -253,4 +253,5 @@ Tox = Source(
     filename=TOX_INI,
     extract=get_tox_ini_python_versions,
     update=update_tox_ini_python_versions,
+    check_pypy_consistency=True,
 )